move color helpers to ui package
