+++ ui/color.go
@@ -1,4 +1,4 @@
-package ioutil
+package ui
 
 import "os"
